Added check for BT 5 features and fixed advertising
None of the advertising features are available on the Moto G5 Plus
It also appears that something is different about how the adv is
handled with newer APIs because I started to get an adv too long
failure. This lead me to an issue with the UUID because Android
expects UUIDs to be of one form, but the established UART UUID
doesn't fit this form. Noted in comments.